% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTAssignment.R
\name{ctassign}
\alias{ctassign}
\title{Clinical Trial Simple Assignment Function}
\usage{
ctassign(data = data.frame(), treatments = c("A", "B"),
  treatmentnames = "Treatment", proportions = c(1, 1))
}
\arguments{
\item{data}{A data frame containing patient information}

\item{treatments}{Name of all possible treatments to assign. Defaults to "A" and "B"}

\item{treatmentnames}{Name of the data frame column where the assigned treatments should be written
Defaults to "Treatment"}

\item{proportions}{A vector containing the relative probability with which eah treatment should be
chosen, when randomly assigned, or how many times each treatment should show in each block, when
blocking. Defaults to c(1,1)}
}
\description{
Assigns patients to treatments in clinical trials.
}
