% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTAssignment.R
\name{ctminimise}
\alias{ctminimise}
\title{Clinical Trial Minimisation Function}
\usage{
ctminimise(data, patients = "Patient", treatmentnames = "Treatment",
  treatments = c("A", "B"), proportion = c(1, 1), factors = NULL)
}
\arguments{
\item{data}{A data frame containing patient information}

\item{patients}{Name of the data frame column with patient identification. Defaults to "Patient"}

\item{treatmentnames}{Name of the data frame column where the assigned treatments should be written
Defaults to "Treatment"}

\item{treatments}{Name of all possible treatments to assign. Defaults to "A" and "B"}

\item{factors}{A vector containing the names of the columns to be taken as factors in minimization
assignment}

\item{proportions}{A vector containing the relative probability with which eah treatment should be
chosen, when randomly assigned, or how many times each treatment should show in each block, when
blocking. Defaults to c(1,1)}
}
\description{
Assigns patients to treatments in clinical trials by minimisation.
}
