% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTAssignment.R
\name{ctsetdataframe}
\alias{ctsetdataframe}
\title{Data frame editing function}
\usage{
ctsetdataframe(data)
}
\arguments{
\item{data}{A data frame containing patient information}
}
\description{
Creates "Patient" and "Treatment" column
}
