% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTAssignment.R
\name{ctstrat}
\alias{ctstrat}
\title{Stratified randomization function}
\usage{
ctstrat(data, strats = NULL, treatments = c("A", "B"),
  proportions = c(1, 1), simple = TRUE, treatmentnames = "Treatment")
}
\arguments{
\item{data}{A data frame containing patient information}

\item{strats}{A vector containing the names of the columns to be used as strata for stratified
assignment}

\item{treatments}{Name of all possible treatments to assign. Defaults to "A" and "B"}

\item{proportions}{A vector containing the relative probability with which eah treatment should be
chosen, when randomly assigned, or how many times each treatment should show in each block, when
blocking. Defaults to c(1,1)}

\item{simple}{Should simple randomization or blocking randomization be used within each stratum?
Defaults to TRUE}

\item{treatmentnames}{Name of the data frame column where the assigned treatments should be written
Defaults to "Treatment"}
}
\description{
Divides patients into strata and assigns treatments by blocking randomization within each stratum
}
