% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTAssignment.R
\name{CTAssignment}
\alias{CTAssignment}
\title{Clinical Trial Treatment Assignment Functions}
\usage{
ctsetdataframe(data)
ctassign = (data = data.frame(), treatments = c("A","B"), treatmentnames = "Treatment", proportions = c(1,1))
ctblock(data = data.frame(), treatments = c("A","B"), proportions = c(1,1), treatmentnames = "Treatment")
ctstrat(data, strats = NULL, treatments = c("A","B"), proportions = c(1,1), simple = TRUE, treatmentnames = "Treatment")
ctminimize(data, patients = "Patient", treatmentnames = "Treatment", treatments = c("A","B"), proportion = c(1,1), factors = NULL)
}
\arguments{
\item{data}{A data frame containing patient information}

\item{patients}{Name of the data frame column with patient identification. Defaults to "Patient"}

\item{treatmentnames}{Name of the data frame column where the assigned treatments should be written
Defaults to "Treatment"}

\item{treatments}{Name of all possible treatments to assign. Defaults to "A" and "B"}

\item{proportions}{A vector containing the relative probability with which eah treatment should be
chosen, when randomly assigned, or how many times each treatment should show in each block, when
blocking. Defaults to c(1,1)}

\item{factors}{A vector containing the names of the columns to be taken as factors in minimization
assignment}

\item{strats}{A vector containing the names of the columns to be used as strata for stratified
assignment}
}
\description{
Assigns patients to treatments in clinical trials.
}
\value{
\item{ctsetdataframe}{creates a "Patient" and "Treatment" column, if they don't already exist}
\item{ctassign}{does simple random assignment}
\item{ctblock}{does blocked randomization}
\item{ctstrat}{does stratified randomization}
\item{ctminimise}{does minimisation}
}
\keyword{
\item{ctassign}
\item{ctminimise}
\item{ctblock}
\item{ctstrat}
\item{ctsetdrataframe}
}
